Modified Planner: M0 move were not taken into acount, and was blocking system.

+++ planner.h
@@ -37,7 +37,7 @@ struct moveBuffer {
 	bool active;												// remember if the buffer is active, i.e. if it has been set or if it's empty
 	bool compute;												// Rememeber if the planner has already calulated the deltas and increments
 
-	int pos[3];													// The position we must go to
+	long pos[3];													// The position we must go to
 	int speed;													// Ditto speed. Applies to the curent movement. Galvo increments/second.
 	char mode;													// Stores the current mode: 0 = fast movement, else = calibrated movement
 
@@ -45,7 +45,7 @@ struct moveBuffer {
 
 	// TODO: SEE WITH TESTS IF deltaTotal HAS TO BEE A DOUBLE INSTEAD OF AN UNSIGNED INT.
 	unsigned int deltaTotal;									// The delta between the start and the final position.
-	int delta[3];													
+	long delta[3];													
 
 	long steps;													// Stores the number ot steps for this move (== (delta / speed) * ISR_FREQUENCY)
 	long nowSteps;
